+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0-beta.4] 2022-11-04
+### Added
+- Created maplibre adapter url tests
+
+### Changed
+- Shortened url template params from `{<name>-source}` to `{s<name>}`
+- Upgraded cypress visual regression testing 
+- Removed package.json type `module` due to cypress issues
+- Improved README.md
+
 ## [1.0.0-beta.3] 2022-11-03
 ### Fixed
 - Conditional package exports